requirements.txt
openai>=1.12.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
